refactor: divided cli from server
